Refactor mobile create button and remove teleport
